[MO] List all countries even if disabled on shop
In a multishop setup, if a country is disabled on a shop it cannot be
selected as the filtered country for customers.
If a merchant in a EU country has a multishop setup where each shop is
designed for a specific country, the merchant cannot choose its own
country to disable VAT reduction.